+++ app/src/main/java/com/example/cinet/ClassDialog.kt
@@ -25,7 +25,8 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import com.example.cinet.com.example.cinet.data.model.CampusRegistry
+import com.example.cinet.data.model.CampusRegistry
+import com.example.cinet.ClassItem
 
 @Composable
 fun ClassDialog(
