+++ app/src/main/java/com/example/cinet/ClassReminderScheduler.kt
@@ -12,6 +12,7 @@ import java.time.LocalTime
 import java.time.ZoneId
 import java.time.format.DateTimeFormatter
 import java.util.Locale
+import com.example.cinet.ClassItem
 
 object ClassReminderScheduler {
