+++ app/src/main/java/com/example/cinet/Map.kt
@@ -66,7 +66,7 @@ import androidx.compose.ui.platform.LocalFocusManager
 import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
-import com.example.cinet.com.example.cinet.data.model.CampusRegistry
+import com.example.cinet.data.model.CampusRegistry
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.maps.CameraUpdateFactory
 import com.google.android.gms.maps.model.BitmapDescriptorFactory
