+++ app/src/main/java/com/example/cinet/data/remote/FirestoreCollections.kt
@@ -1,4 +1,4 @@
-package com.example.cinet.data
+package com.example.cinet.data.remote
 
 object FirestoreCollections {
     const val USERS = "users"
