+++ app/src/main/java/com/example/cinet/data/remote/FirestoreRepository.kt
@@ -1,7 +1,7 @@
 package com.example.cinet.data.remote
 
 import android.util.Log
-import com.example.cinet.data.FirestoreCollections
+import com.example.cinet.data.remote.FirestoreCollections
 import com.example.cinet.data.model.CampusEvent
 import com.example.cinet.data.model.UserProfile
 import com.google.firebase.Timestamp
